Fix pg_autoctl status when pg_autoctl is not running.
And when it's running and --json is used, match the same format as in the
command pg_autoctl show file --pid --contents --json, adding the status of
the pids (running, or stale).

+++ src/bin/pg_autoctl/cli_service.c
@@ -424,28 +424,33 @@ cli_service_status(int argc, char **argv)
 
 	if (!file_exists(pathnames->pid))
 	{
-		log_info("pg_autoctl pid file \"%s\" does not exists", pathnames->pid);
+		log_debug("pg_autoctl pid file \"%s\" does not exists", pathnames->pid);
 
 		if (pg_setup_is_running(pgSetup))
 		{
 			log_fatal("Postgres is running at \"%s\" with pid %d",
 					  pgSetup->pgdata, pgSetup->pidFile.pid);
-			exit(EXIT_CODE_INTERNAL_ERROR);
 		}
+
+		log_info("pg_autoctl is not running at \"%s\"", pgSetup->pgdata);
+		exit(PG_CTL_STATUS_NOT_RUNNING);
 	}
 
 	/* ok now we have a pidfile for pg_autoctl */
-	if (read_pidfile(pathnames->pid, &pid))
+	if (read_pidfile(pathnames->pid, &pid) && pid > 0)
 	{
 		if (kill(pid, 0) != 0)
 		{
 			log_error("pg_autoctl pid file contains stale pid %d", pid);
-			exit(EXIT_CODE_INTERNAL_ERROR);
+			exit(PG_CTL_STATUS_NOT_RUNNING);
 		}
 	}
 
 	/* and now we know pg_autoctl is running */
-	log_info("pg_autoctl is running with pid %d", pid);
+	if (pid > 0)
+	{
+		log_info("pg_autoctl is running with pid %d", pid);
+	}
 
 	/* add a word about the Postgres service itself */
 	if (pg_setup_is_ready(pgSetup, false))
@@ -460,29 +465,8 @@ cli_service_status(int argc, char **argv)
 
 	if (outputJSON)
 	{
-		JSON_Value *js = json_value_init_object();
-		JSON_Value *jsPostgres = json_value_init_object();
+		bool includeStatus = true;
 
-		JSON_Value *jsPGAutoCtl = json_value_init_object();
-		JSON_Object *jsobj = json_value_get_object(jsPGAutoCtl);
-
-		JSON_Object *root = json_value_get_object(js);
-
-		/* prepare both JSON objects */
-		json_object_set_number(jsobj, "pid", (double) pid);
-
-		if (!pg_setup_as_json(pgSetup, jsPostgres))
-		{
-			/* can't happen */
-			exit(EXIT_CODE_INTERNAL_ERROR);
-		}
-
-		/* concatenate JSON objects into a container object */
-		json_object_set_value(root, "postgres", jsPostgres);
-		json_object_set_value(root, "pg_autoctl", jsPGAutoCtl);
-
-		(void) cli_pprint_json(js);
+		(void) fprint_pidfile_as_json(pathnames->pid, includeStatus);
 	}
-
-	exit(EXIT_CODE_QUIT);
 }
